[bug fix] move with negative number [change] keyboard on Robot to PySimbotMap widget [add] shortkey for change food position
