+++ pysimbotlib/Window.py
@@ -19,6 +19,8 @@ from random import randint
 from kivy.logger import Logger
 from .scaler import Scaler
 
+import random
+
 from . import Obstacle
 from . import Objective
 
@@ -35,6 +37,8 @@ class PySimbotMap(Widget):
 
     def __init__(self, mapPath, **kwargs):
         super(PySimbotMap, self).__init__(**kwargs)
+        self._keyboard = Window.request_keyboard(self._keyboard_closed, self)
+        self._keyboard.bind(on_key_down=self._on_keyboard_down)
         self.mapPath = mapPath
         self.robots = list()
     
@@ -57,7 +61,44 @@ class PySimbotMap(Widget):
         for robot in self.robots:
             robot.update()
         self.after_update()
+    
+    def _keyboard_closed(self):
+        self._keyboard.unbind(on_key_down=self._on_keyboard_down)
+        self._keyboard = None
+    
+    def _on_keyboard_down(self, keyboard, keycode, text, modifiers):
+        if keycode[1] == 'w':
+            for r in self.robots:
+                r.move(5)
+        elif keycode[1] == 'a':
+            for r in self.robots:
+                r.turn(-5)
+        elif keycode[1] == 'd':
+            for r in self.robots:
+                r.turn(5)
+        elif keycode[1] == 's':
+            for r in self.robots:
+                r.move(-5)
+        elif keycode[1] == 'n':
+            for obj in self.objectives.get_objectives():
+                new_pos = Vector(random.randrange(self.playground.size[0]), random.randrange(self.playground.size[1]))
+                while not self.isFoodPosValid(obj, new_pos):
+                    new_pos = Vector(random.randrange(self.playground.size[0]), random.randrange(self.playground.size[1]))
+                obj.pos = new_pos
 
+    def isFoodPosValid(self, obj, pos):
+        # check wall
+        if pos.x < 0 or pos.x > self.size[0] - obj.size[0]:
+            return False
+        if pos.y < 0 or pos.y > self.size[1] - obj.size[1]:
+            return False
+
+        # check obstracles
+        for obs in self.obstacles.get_obstacles():
+            if (obs.pos[0] < pos.x < obs.pos[0] + obs.size[0] or obs.pos[0] < pos.x + obj.size[0] < obs.pos[0] + obs.size[0])\
+                and (obs.pos[1] < pos.y < obs.pos[1] + obs.size[1] or obs.pos[1] < pos.y + obj.size[1] < obs.pos[1] + obs.size[1]):
+                return False
+        return True
 
 class PySimbotApp(App):
     title = 'PySimbot (Created by Saran(Bird) Khotsathian CPE25)'
@@ -73,7 +114,7 @@ class PySimbotApp(App):
         self.simbotMap = None
     
     def build(self):
-        Window.size = (700, 600)
+        Window.size = (900, 600)
         obmap = Builder.load_file(self.mapPath)
         robot = Builder.load_file('pysimbotlib/widget/robot.kv')
         sbmap = Builder.load_file('pysimbotlib/widget/simbotmap.kv')
