+++ pysimbotlib/widget/simbotmap.kv
@@ -33,7 +33,7 @@
 <PySimbotMap>:
     playground: playground
     pos: 0, 0
-    size: 800, 600
+    size: 900, 600
     canvas:
         Color:
             rgba: 0.7, 0.7, 0.7, 1
@@ -41,31 +41,44 @@
             pos: self.pos
             size: self.size
 
-    RelativeLayout:
+    BoxLayout:
+        orientation: 'horizontal'
         size_hint: None, None
-        size: 700, 600
-        pos: 0, 0 
-        canvas:
-            Color:
-                rgba: 0,0,0,0.7
-            Rectangle:
-                pos: self.pos
-                size: self.size
+        size: 900, 600
 
-            Color:
-                rgba: 1,1,1,0.7
-            Rectangle:
-                size: self.width - 10, self.height - 10
-                pos: self.x + 5, self.y + 5
-        
-        Widget:
-            id: playground
+        RelativeLayout:
             size_hint: None, None
-            size: self.parent.width - 10, self.parent.height - 10
-            pos: self.parent.x + 5, self.parent.y + 5
+            size: 700, 600
+            pos: 0, 0 
             canvas:
                 Color:
-                    rgba: 1, 0.78, 0.78, 0.5
+                    rgba: 0,0,0,0.7
                 Rectangle:
                     pos: self.pos
                     size: self.size
+
+                Color:
+                    rgba: 1,1,1,0.7
+                Rectangle:
+                    size: self.width - 10, self.height - 10
+                    pos: self.x + 5, self.y + 5
+            
+            Widget:
+                id: playground
+                size_hint: None, None
+                size: self.parent.width - 10, self.parent.height - 10
+                pos: self.parent.x + 5, self.parent.y + 5
+                canvas:
+                    Color:
+                        rgba: 1, 0.78, 0.78, 0.5
+                    Rectangle:
+                        pos: self.pos
+                        size: self.size
+    
+        Label:
+            id: 'lbl_description'
+            text: 'Shortcuts\nw,a,s,d: move robot\nn: change food pos'
+            text_size: self.size
+            padding: (15, 10)
+            halign: 'left'
+            valign: 'top'
